requirements.txt
dotenv~=0.9.9
python-dotenv~=1.1.0
loguru~=0.7.3
requests~=2.32.3
feedparser~=6.0.11
newspaper3k~=0.2.8
openai~=1.68.2
DateTime~=5.5